# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/employee/employeeCompleteData": {
            "post": {
                "tags": [
                    "Employee"
                ],
                "summary": "Add details to employee",
                "operationId": "1dbb2ce22965cc6ef99407331b317302",
                "requestBody": {
                    "required": false,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "description": {
                                        "description": "Description of the employee",
                                        "type": "string"
                                    },
                                    "imageSSN": {
                                        "description": "صورة الباطاقة",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "livePhoto": {
                                        "description": "صورة لايف",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "nationalId": {
                                        "description": " الرقم القومي",
                                        "type": "string"
                                    },
                                    "phone_number_2": {
                                        "description": "phone_number_2",
                                        "type": "string"
                                    },
                                    "mobile_number_1": {
                                        "description": "mobile_number_1",
                                        "type": "string"
                                    },
                                    "mobile_number_2": {
                                        "description": "mobile_number_2",
                                        "type": "string"
                                    },
                                    "fax_number": {
                                        "description": "fax_number",
                                        "type": "string"
                                    },
                                    "whatsapp_number": {
                                        "description": "whatsapp_number",
                                        "type": "string"
                                    },
                                    "facebook_link": {
                                        "description": "facebook_link",
                                        "type": "string"
                                    },
                                    "website": {
                                        "description": "website",
                                        "type": "string"
                                    },
                                    "city": {
                                        "description": "city",
                                        "type": "string"
                                    },
                                    "bitTitle": {
                                        "description": "bitTitle",
                                        "type": "string"
                                    },
                                    "street": {
                                        "description": "street",
                                        "type": "string"
                                    },
                                    "specialMarque": {
                                        "description": "specialMarque",
                                        "type": "string"
                                    },
                                    "zipCode": {
                                        "description": " city zip code",
                                        "type": "string"
                                    },
                                    "lat": {
                                        "description": "lat",
                                        "type": "integer"
                                    },
                                    "long": {
                                        "description": "long",
                                        "type": "integer"
                                    },
                                    "user_id": {
                                        "description": "User ID",
                                        "type": "integer"
                                    },
                                    "service_id": {
                                        "description": "Service ID",
                                        "type": "integer"
                                    },
                                    "section_id": {
                                        "description": "Section ID",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Employee details add successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "successfully"
                                        },
                                        "employee": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/employee/employeeProfile/{id}": {
            "get": {
                "tags": [
                    "Employee"
                ],
                "summary": "Show employee profile",
                "description": "Show employee profile by user id where type employee",
                "operationId": "2297a9c817dc43c2ad915af565cd1fe9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user where user type employee to show profile",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show employee profile successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Employee not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/employee/section/{section_id}/service/{service_id}": {
            "get": {
                "tags": [
                    "Employee"
                ],
                "summary": "Get all employees in section and service",
                "operationId": "8cf253a0b16580eabe231c90d4f50577",
                "parameters": [
                    {
                        "name": "section_id",
                        "in": "path",
                        "description": "ID of the section",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "service_id",
                        "in": "path",
                        "description": "ID of the service",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "allemployee": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No employees found in this service",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No employee found in this service"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/employee/showAllEmployeeBylocation/{city}": {
            "post": {
                "tags": [
                    "Employee"
                ],
                "description": "Show all employee by city",
                "operationId": "aeb8198ae718b4b5177093088677b637",
                "parameters": [
                    {
                        "name": "city",
                        "in": "path",
                        "description": "city to show all employees in this city",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show all employees successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Employee not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}